Added support for SAP model potentials and updated users-guide.tex accordingly.
Added an error message in inputData.f90 when a model potential is defined but
the method selected is not OED one.

Added 2 style files in docs to help compile the documentation.

Fixed a glitch in writeDisk which caused the program to break when retrieving
orbitals with "orbpot noexch" card present.